+++ CHANGELOG.md
@@ -6,7 +6,7 @@ This means:
 * A change in the minor version implies a new feature, without breaking compatibility.
 * A change in the patch version implies a bug fix.
 
-## [Unreleased]
+## [1.1.0] - 2022-04-18
 
 ### Changed
 - Better support for longer running test suites:
@@ -14,6 +14,8 @@ This means:
       before showing the output
     - ignore file saves while a test is already running
 - Clear the output channel before each run
+- Add a separate icon for errors during the test run
+- Add configuration options to map exit codes to status icons
 
 ## [1.0.0] - 2021-11-21